backend/requirements.txt
anthropic==0.74.1
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
flask==3.0.0
flask-cors==4.0.0
playwright==1.56.0
websockets==15.0.1
lightpanda==1.0.0

